refactor: move UI to layout file so routes / and /thread/:id share it
